requirements.txt
langgraph
langchain
langchain-community
langchain-core
duckduckgo-search
ddgs
groq
streamlit
python-dotenv